priors reorganize & tests
