+++ sparsecoding/priors/__init__.py
@@ -1,0 +1,9 @@
+from .l0_prior import L0Prior
+from .prior import Prior
+from .spike_slab_prior import SpikeSlabPrior
+
+__all__ = [
+  "Prior",
+  "L0Prior",
+  "SpikeSlabPrior",
+]
